tipuskonverzio

+++ java_3/src/java_3/Main.java
@@ -9,7 +9,7 @@ public class Main {
 		int osszeg=a+b;
 		int kulonbseg=a-b;
 		int szorzat=a*b;
-		double hanyados=a/b;
+		double hanyados=(double)a/b;
 		
 		
 		System.out.println(osszeg);
